+++ src/js/player.js
@@ -1,0 +1,5 @@
+export const Player = (askedName, askedCoin) => {
+  const name = askedName;
+  const coin = askedCoin;
+  return { name, coin };
+};
